# core-lending-regulatory-compliance-sdk/target/classes/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-regulatory-compliance
  description: Lending Regulatory & Compliance Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/loan-compliance
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/aml-cases/{amlCaseId}:
    get:
      tags:
        - AmlCase
      summary: Get an AML case by ID
      operationId: getById_3
      parameters:
        - name: amlCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AmlCaseDTO'
    put:
      tags:
        - AmlCase
      summary: Update an AML case
      operationId: update_3
      parameters:
        - name: amlCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AmlCaseDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AmlCaseDTO'
    delete:
      tags:
        - AmlCase
      summary: Delete an AML case
      operationId: delete_3
      parameters:
        - name: amlCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/aml-cases/{amlCaseId}/actions/{amlActionId}:
    get:
      tags:
        - AmlAction
      summary: Get an AML action by ID
      operationId: getById_5
      parameters:
        - name: amlCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: amlActionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AmlActionDTO'
    put:
      tags:
        - AmlAction
      summary: Update an AML action
      operationId: update_5
      parameters:
        - name: amlCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: amlActionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AmlActionDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AmlActionDTO'
    delete:
      tags:
        - AmlAction
      summary: Delete an AML action
      operationId: delete_5
      parameters:
        - name: amlCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: amlActionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/aml-cases:
    get:
      tags:
        - AmlCase
      summary: List or search AML cases
      operationId: findAll_3
      parameters:
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.reportingRunId
          in: query
          description: Exact filter for reporting run id
          required: false
          schema:
            type: string
        - name: filters.customerId
          in: query
          description: Exact filter for customer id
          required: false
          schema:
            type: string
        - name: filters.transactionId
          in: query
          description: Exact filter for transaction id
          required: false
          schema:
            type: string
        - name: filters.amlCaseStatus
          in: query
          description: Exact filter for aml case status
          required: false
          schema:
            type: string
            enum:
              - OPEN
              - INVESTIGATING
              - REPORTED
              - CLOSED
              - ESCALATED
        - name: filters.caseOpenedAt
          in: query
          description: Exact filter for case opened at
          required: false
          schema:
            type: string
        - name: filters.caseClosedAt
          in: query
          description: Exact filter for case closed at
          required: false
          schema:
            type: string
        - name: filters.riskLevel
          in: query
          description: Exact filter for risk level
          required: false
          schema:
            type: string
            enum:
              - LOW
              - MEDIUM
              - HIGH
              - CRITICAL
        - name: filters.notes
          in: query
          description: Exact filter for notes
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseAmlCaseDTO'
    post:
      tags:
        - AmlCase
      summary: Create a new AML case
      operationId: create_3
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AmlCaseDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AmlCaseDTO'
...
  /api/v1/aml-cases/{amlCaseId}/actions:
    get:
      tags:
        - AmlAction
      summary: List or search AML actions for a case
      operationId: findAll_5
      parameters:
        - name: amlCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.amlCaseId
          in: query
          description: Exact filter for aml case id
          required: false
          schema:
            type: string
        - name: filters.actionDate
          in: query
          description: Exact filter for action date
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[actionDate].from
          in: query
          description: Filter action date from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[actionDate].to
          in: query
          description: Filter action date to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.actionType
          in: query
          description: Exact filter for action type
          required: false
          schema:
            type: string
            enum:
              - REVIEW
              - CUSTOMER_CONTACT
              - REPORT_TO_AUTHORITY
              - ACCOUNT_FREEZE
              - DOCUMENT_REQUEST
        - name: filters.outcome
          in: query
          description: Exact filter for outcome
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseAmlActionDTO'
    post:
      tags:
        - AmlAction
      summary: Create a new AML action
      operationId: create_5
      parameters:
        - name: amlCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AmlActionDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AmlActionDTO'
components:
  schemas:
...
    AmlCaseDTO:
      required:
        - amlCaseStatus
        - caseOpenedAt
        - customerId
        - reportingRunId
        - riskLevel
        - transactionId
      type: object
      properties:
        amlCaseId:
          type: string
          format: uuid
          readOnly: true
        reportingRunId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        transactionId:
          type: string
          format: uuid
        amlCaseStatus:
          type: string
          enum:
            - OPEN
            - INVESTIGATING
            - REPORTED
            - CLOSED
            - ESCALATED
        caseOpenedAt:
          type: string
          format: date
        caseClosedAt:
          type: string
          format: date
        riskLevel:
          type: string
          enum:
            - LOW
            - MEDIUM
            - HIGH
            - CRITICAL
        notes:
          maxLength: 2000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AmlActionDTO:
      required:
        - actionDate
        - actionType
        - amlCaseId
        - outcome
      type: object
      properties:
        amlActionId:
          type: string
          format: uuid
          readOnly: true
        amlCaseId:
          type: string
          format: uuid
        actionDate:
          type: string
          format: date-time
        actionType:
          type: string
          enum:
            - REVIEW
            - CUSTOMER_CONTACT
            - REPORT_TO_AUTHORITY
            - ACCOUNT_FREEZE
            - DOCUMENT_REQUEST
        outcome:
          maxLength: 1000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaginationResponseAmlCaseDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/AmlCaseDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...
    PaginationResponseAmlActionDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/AmlActionDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.