# core-lending-regulatory-compliance-web/src/main/java/com/firefly/core/lending/compliance/web/controllers/aml/v1/AmlActionController.java
/*
 * Copyright 2025 Firefly Software Foundation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */


package com.firefly.core.lending.compliance.web.controllers.aml.v1;

import org.fireflyframework.core.filters.FilterRequest;
import org.fireflyframework.core.queries.PaginationResponse;
import com.firefly.core.lending.compliance.core.services.aml.v1.AmlActionService;
import com.firefly.core.lending.compliance.interfaces.dtos.aml.v1.AmlActionDTO;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.tags.Tag;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import reactor.core.publisher.Mono;

import java.util.UUID;

@RestController
@RequestMapping("/api/v1/aml-cases/{amlCaseId}/actions")
@Tag(name = "AmlAction", description = "Operations on actions within an AML case")
@RequiredArgsConstructor
public class AmlActionController {

    private final AmlActionService service;

    @GetMapping
    @Operation(summary = "List or search AML actions for a case")
    public Mono<ResponseEntity<PaginationResponse<AmlActionDTO>>> findAll(
            @PathVariable UUID amlCaseId,
            @ModelAttribute FilterRequest<AmlActionDTO> filterRequest) {

        return service.findAll(amlCaseId, filterRequest)
                .map(ResponseEntity::ok);
    }

    @PostMapping
    @Operation(summary = "Create a new AML action")
    public Mono<ResponseEntity<AmlActionDTO>> create(
            @PathVariable UUID amlCaseId,
            @RequestBody AmlActionDTO dto) {

        return service.create(amlCaseId, dto)
                .map(ResponseEntity::ok);
    }

    @GetMapping("/{amlActionId}")
    @Operation(summary = "Get an AML action by ID")
    public Mono<ResponseEntity<AmlActionDTO>> getById(
            @PathVariable UUID amlCaseId,
            @PathVariable UUID amlActionId) {

        return service.getById(amlCaseId, amlActionId)
                .map(ResponseEntity::ok);
    }

    @PutMapping("/{amlActionId}")
    @Operation(summary = "Update an AML action")
    public Mono<ResponseEntity<AmlActionDTO>> update(
            @PathVariable UUID amlCaseId,
            @PathVariable UUID amlActionId,
            @RequestBody AmlActionDTO dto) {

        return service.update(amlCaseId, amlActionId, dto)
                .map(ResponseEntity::ok);
    }

    @DeleteMapping("/{amlActionId}")
    @Operation(summary = "Delete an AML action")
    public Mono<ResponseEntity<Void>> delete(
            @PathVariable UUID amlCaseId,
            @PathVariable UUID amlActionId) {

        return service.delete(amlCaseId, amlActionId)
                .thenReturn(ResponseEntity.noContent().build());
    }
}